Move the application modules
