+++ src/amulet/app/_splash/__init__.py
@@ -1,0 +1,1 @@
+from ._splash import Splash
