+++ src/amulet/app/_splash/_splash_ui.ui
